accrual type fix

+++ internal/model/model.go
@@ -25,8 +25,8 @@ type WithdrawRequest struct {
 }
 
 type WithdrawResponse struct {
-	Order        string    `json:"order"`
-	Sum          float64   `json:"sum"`
+	Order       string    `json:"order"`
+	Sum         float64   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
 
@@ -35,7 +35,7 @@ type Sum struct {
 }
 
 type AccrualResponse struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual,omitempty"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual,omitempty"`
 }
